Extract components & improve naming
