+++ src/app/core/components/bread-crumb/bread-crumb.component.spec.ts
@@ -2,7 +2,7 @@ import {BreadCrumbComponent} from './bread-crumb.component';
 import {ComponentFixture} from '@angular/core/testing/src/component_fixture';
 import {fakeAsync, TestBed, tick} from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
-import {ErrorLogService, XmppService} from '../';
+import {ErrorLogService, XmppService} from '../../index';
 import {Router} from '@angular/router';
 import SpyObj = jasmine.SpyObj;
 import createSpyObj = jasmine.createSpyObj;
