+++ src/app/core/components/error-page/error-page.component.spec.ts
@@ -1,4 +1,4 @@
-import {ErrorPageComponent} from '.';
+import {ErrorPageComponent} from './index';
 import {fakeAsync, TestBed} from '@angular/core/testing';
 import {ComponentFixture} from '@angular/core/testing/src/component_fixture';
 import {ActivatedRoute} from '@angular/router';
