+++ src/app/core/components/header/header.component.spec.ts
@@ -1,7 +1,7 @@
 import {TestBed} from '@angular/core/testing';
 import {ComponentFixture} from '@angular/core/testing/src/component_fixture';
 import {RouterTestingModule} from '@angular/router/testing';
-import {HeaderComponent} from '.';
+import {HeaderComponent} from './index';
 
 
 describe('HeaderComponent', () => {
