+++ src/app/core/components/index.ts
@@ -1,0 +1,3 @@
+export * from './bread-crumb';
+export * from './error-page';
+export * from './header';
