+++ src/app/core/feature-detection/index.ts
@@ -1,0 +1,2 @@
+export * from './feature.service';
+export * from './feature-detection-guard.service';
