+++ src/app/core/index.ts
@@ -1,10 +1,8 @@
-export * from './bread-crumb';
-export * from './error-page';
+export * from './components';
+export * from './config-loader';
 export * from './errors';
-export * from './header';
+export * from './feature-detection';
+export * from './navigation';
 export * from './notifications';
-export * from './xmpp';
-export * from './xmpp-feature-detection';
-export * from './config-loader/';
 export * from './utils/';
-export * from './navigation/';
+export * from './xmpp';
