--- src/app/core/xmpp-feature-detection/index.ts
@@ -1,2 +1,0 @@
-export * from './xmpp-feature.service';
-export * from './xmpp-feature-guard.service';
